requirements.txt
Flask==3.1.0
Flask_Login==0.6.3
flask_mail==0.10.0
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
python-dotenv==1.1.0
SQLAlchemy==2.0.30
Werkzeug==3.1.3
WTForms==3.2.1
